cria estrutura base do projeto
