+++ routes/__init__.py
@@ -1,0 +1,2 @@
+from fastapi import APIRouter;
+router = APIRouter()
