+++ routes/status.py
@@ -1,0 +1,5 @@
+from . import router;
+
+@router.get("/")
+def status():
+    return{"status": "ok"}
